Move lintablePaths impl into Validator
