+++ Sources/IBLinterKit/Validator.swift
@@ -19,7 +19,7 @@ public class Validator {
     public func validate(workDirectory: URL, config: Config) -> [Violation] {
         let context = Context(config: config, workDirectory: workDirectory, externalRules: externalRules)
         let rules = Rules.rules(context)
-        let (xibs, storyboards) = config.lintablePaths(workDirectory: workDirectory)
+        let (xibs, storyboards) = lintablePaths(workDirectory: workDirectory, config: config)
         return validateXib(files: xibs, rules: rules, config: config)
             + validateStoryboard(files: storyboards, rules: rules, config: config)
     }
@@ -52,6 +52,66 @@ public class Validator {
                 }
             }
         }
+    }
+
+    private struct InterfaceBuilderFiles {
+        var xibPaths: Set<URL> = []
+        var storyboardPaths: Set<URL> = []
+    }
+
+    private final class LintableFileMatcher {
+        let fileManager: FileManager
+        let globber: Glob
+        init(fileManager: FileManager = .default) {
+            self.fileManager = fileManager
+            self.globber = Glob(fileManager: fileManager)
+        }
+
+        func interfaceBuilderFiles(withPatterns patterns: [URL]) -> InterfaceBuilderFiles {
+            return patterns.flatMap { globber.expandRecursiveStars(pattern: $0.path) }
+                .reduce(into: InterfaceBuilderFiles()) { result, path in
+                    let files = self.interfaceBuilderFiles(atPath: URL(fileURLWithPath: path))
+                    result.xibPaths.formUnion(files.xibPaths)
+                    result.storyboardPaths.formUnion(files.storyboardPaths)
+            }
+        }
+
+        func interfaceBuilderFiles(atPath path: URL) -> InterfaceBuilderFiles {
+            var xibs: Set<URL> = []
+            var storyboards: Set<URL> = []
+            guard let enumerator = fileManager.enumerator(at: path, includingPropertiesForKeys: [.isRegularFileKey]) else {
+                return InterfaceBuilderFiles(xibPaths: xibs, storyboardPaths: storyboards)
+            }
+
+            for element in enumerator {
+                guard let absolute = element as? URL else { continue }
+                switch absolute.pathExtension {
+                case "xib": xibs.insert(absolute)
+                case "storyboard": storyboards.insert(absolute)
+                default: continue
+                }
+            }
+            return InterfaceBuilderFiles(xibPaths: xibs, storyboardPaths: storyboards)
+        }
+    }
+
+    public func lintablePaths(workDirectory: URL, config: Config, fileManager: FileManager = .default) -> (xib: Set<URL>, storyboard: Set<URL>) {
+        let matcher = LintableFileMatcher(fileManager: fileManager)
+        let files: InterfaceBuilderFiles
+        if config.included.isEmpty {
+            files = matcher.interfaceBuilderFiles(atPath: workDirectory)
+        } else {
+            files = matcher.interfaceBuilderFiles(
+                withPatterns: config.included.map { workDirectory.appendingPathComponent($0) }
+            )
+        }
+
+        let excluded = matcher.interfaceBuilderFiles(
+            withPatterns: config.excluded.map { workDirectory.appendingPathComponent($0) }
+        )
+        let xibLintablePaths = files.xibPaths.subtracting(excluded.xibPaths)
+        let storyboardLintablePaths = files.storyboardPaths.subtracting(excluded.storyboardPaths)
+        return (xibLintablePaths, storyboardLintablePaths)
     }
 }
