+++ Tests/IBLinterKitTest/Utils/Config+LintablePathsTests.swift
@@ -2,7 +2,7 @@
 import XCTest
 
 
-class ConfigLintablePathsTests: XCTestCase {
+class LintablePathsTests: XCTestCase {
 
     let fixture = Fixture()
 
@@ -12,8 +12,9 @@ class ConfigLintablePathsTests: XCTestCase {
             excluded: ["Level1_1"], included: [],
             customModuleRule: [], baseClassRule: [], reporter: ""
         )
+        let validator = Validator(externalRules: [])
         let projectPath = fixture.path("Resources/Utils/Glob/ProjectMock")
-        let lintablePaths = config.lintablePaths(workDirectory: projectPath).xib
+        let lintablePaths = validator.lintablePaths(workDirectory: projectPath, config: config).xib
 
         XCTAssertEqual(
             lintablePaths.map { $0.path },
@@ -27,8 +28,9 @@ class ConfigLintablePathsTests: XCTestCase {
             excluded: [], included: ["Level1_2"],
             customModuleRule: [], baseClassRule: [], reporter: ""
         )
+        let validator = Validator(externalRules: [])
         let projectPath = fixture.path("Resources/Utils/Glob/ProjectMock")
-        let lintablePaths = config.lintablePaths(workDirectory: projectPath).xib
+        let lintablePaths = validator.lintablePaths(workDirectory: projectPath, config: config).xib
 
         XCTAssertEqual(
             lintablePaths.map { $0.path },
@@ -42,8 +44,9 @@ class ConfigLintablePathsTests: XCTestCase {
             excluded: ["Level1_1"], included: ["Level1_1/Level2_1"],
             customModuleRule: [], baseClassRule: [], reporter: ""
         )
+        let validator = Validator(externalRules: [])
         let projectPath = fixture.path("Resources/Utils/Glob/ProjectMock")
-        let lintablePaths = config.lintablePaths(workDirectory: projectPath).xib
+        let lintablePaths = validator.lintablePaths(workDirectory: projectPath, config: config).xib
 
         XCTAssertEqual(lintablePaths, [])
     }
